[refactor] post 컴포턴트 추출
